+++ actors/power/src/types.rs
@@ -20,8 +20,8 @@ pub const SECTOR_TERMINATION_MANUAL: SectorTermination = 1;
 pub const SECTOR_TERMINATION_FAULTY: SectorTermination = 3;
 
 pub const CRON_QUEUE_HAMT_BITWIDTH: u32 = 6;
-pub const CRON_QUEUE_AMT_BITWIDTH: usize = 6;
-pub const PROOF_VALIDATION_BATCH_AMT_BITWIDTH: usize = 4;
+pub const CRON_QUEUE_AMT_BITWIDTH: u32 = 6;
+pub const PROOF_VALIDATION_BATCH_AMT_BITWIDTH: u32 = 4;
 
 #[derive(Serialize_tuple, Deserialize_tuple)]
 pub struct CreateMinerParams {
